+++ src/App.js
@@ -1,18 +1,22 @@
 import React, {Component} from 'react';
 import logo from './lollipop.svg';
 import './App.css';
-import Banner from './General/banner';
+import BannerTop from './General/bannerTop';
+import BannerBottom from './General/bannerBottom';
+import { thisExpression } from '@babel/types';
 
 
 class App extends Component{
   state = {
-    banner : Banner
+    bannerTop : BannerTop,
+    bannerBottom : BannerBottom
   }
   render() {
     const saludo = saludar();
     return (
     <div className="App">
-      <this.state.banner />
+      <this.state.bannerTop />
+      <this.state.bannerBottom />
       <header className="App-header">
         <p className="logo">EVENTRINET</p>
         <p>Gestionador de eventos academicos</p>
